crc: widen size parameter from uint32_t to size_t
Callers like hints.c and device_id.c pass strlen() which returns
size_t, causing implicit truncation on 64-bit.  Also fix the
alignment mask in the non-x86_64 path to work with size_t.

+++ lib/misc/crc.h
@@ -20,6 +20,6 @@
 
 #define INITIAL_CRC 0xf597a6cf
 
-uint32_t calc_crc(uint32_t initial, const uint8_t *buf, uint32_t size);
+uint32_t calc_crc(uint32_t initial, const uint8_t *buf, size_t size);
 
 #endif
